dataservice/personal_stats/requirements.txt
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
apig-wsgi==2.13.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
apig-wsgi==2.13.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
apig-wsgi==2.13.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
apig-wsgi==2.13.0
boto3==1.21.12
botocore==1.24.12
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
